Sonar inspired cleanup

+++ agent/it-harness/src/main/java/org/glowroot/agent/it/harness/Threads.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2016 the original author or authors.
+ * Copyright 2011-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -129,7 +129,7 @@ public class Threads {
     }
 
     // try to handle under- and over- sleeping for tests that depend on more accurate sleep timing
-    public static void moreAccurateSleep(int millis) throws InterruptedException {
+    public static void moreAccurateSleep(long millis) throws InterruptedException {
         Stopwatch stopwatch = Stopwatch.createStarted();
         if (millis > 10) {
             Thread.sleep(millis - 10);
